services/platform/requirements/dev.txt
# Development requirements for PRAHO Platform
# Include base requirements plus development tools

-r base.txt

# Development tools
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
factory-boy>=3.3.0

# Code quality
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
django-stubs[compatible-mypy]>=4.2.0  # Django type stubs
types-requests>=2.31.0  # Type stubs for requests
types-python-dateutil>=2.8.0  # Type stubs for python-dateutil
pre-commit>=3.4.0

# Documentation
Sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Testing utilities
freezegun>=1.2.0
responses>=0.23.0
mixer>=7.2.0
tblib>=3.1.0  # Better traceback handling in parallel tests

# Database testing
# pytest-postgresql>=5.0.0  # Requires psycopg3, conflicts with psycopg2-binary
# Use Django's built-in TestCase for database tests instead

# Performance profiling
django-silk>=5.0.0
memory-profiler>=0.61.0

# Debug tools
django-debug-toolbar>=4.2.0
django-extensions>=3.2.0
django-ratelimit>=4.1.0  # Rate limiting for password reset
werkzeug>=2.3.0  # For runserver_plus

# Romanian testing data
faker>=19.3.0  # Generate Romanian test data
